+++ src/Elrond/Transactions.cpp
@@ -4,7 +4,7 @@
 // terms governing use, modification, and redistribution, is contained in the
 // file LICENSE at the root of the source code distribution tree.
 
-#include "TransactionFactory.h"
+#include "Transactions.h"
 
 #include "Codec.h"
 #include "HexCoding.h"
